+++ src/section/index.js
@@ -16,9 +16,10 @@ const parse = {
 const section_reg = /[\n^](={1,5}[^=]{1,200}?={1,5})/g;
 
 const doSection = function(section, wiki, options) {
-  wiki = parseTemplates(section, wiki, options);
   // //parse the <ref></ref> tags
   wiki = parse.references(section, wiki, options);
+  //parse-out all {{templates}}
+  wiki = parseTemplates(section, wiki, options);
   // //parse the tables
   wiki = parse.table(section, wiki);
   // //parse the lists
@@ -45,7 +46,8 @@ const splitSections = function(wiki, options) {
     let content = split[i] || '';
     let section = {
       title: '',
-      depth: null
+      depth: null,
+      templates: []
     };
     //figure-out title/depth
     section = parse.heading(section, heading);