+++ src/section/templates/citation.js
@@ -1,9 +1,9 @@
 //grab the {{cite web}} templates, etc
-const parseCitation = function(str, wiki, r, options) {
+const parseCitation = function(str, options) {
   //remove it from main
-  wiki = wiki.replace(str, '');
+  // wiki = wiki.replace(str, '');
   if (options.citations === false) {
-    return wiki;
+    return null;
   }
   //trim start {{ and
   //trim end }}
@@ -15,15 +15,12 @@ const parseCitation = function(str, wiki, r, options) {
   //first line is 'cite web'
   let type = lines[0].match(/cite ([a-z_]+)/i) || [];
   if (type[1]) {
-    obj.cite = type[1] || null;
+    obj.type = type[1] || null;
   }
   for (let i = 1; i < lines.length; i += 1) {
     let arr = lines[i].split(/=/);
     let key = arr[0].trim();
-    let val = arr
-      .slice(1, arr.length)
-      .join('=')
-      .trim();
+    let val = arr.slice(1, arr.length).join('=').trim();
     if (key && val) {
       //turn numbers into numbers
       if (/^[0-9.]+$/.test(val)) {
@@ -32,10 +29,12 @@ const parseCitation = function(str, wiki, r, options) {
       obj[key] = val;
     }
   }
-  if (Object.keys(obj).length > 0) {
-    r.citations = r.citations || [];
-    r.citations.push(obj);
+  if (Object.keys(obj).length === 0) {
+    return null;
   }
-  return wiki;
+  return {
+    template: 'citation',
+    data: obj
+  };
 };
 module.exports = parseCitation;