+++ src/section/templates/index.js
@@ -1,61 +1,68 @@
 const i18n = require('../../data/i18n');
 const findRecursive = require('../../lib/recursive_match');
 const parseInfobox = require('../../infobox/parse-infobox');
-const parseCitation = require('./citation');
-const Infobox = require('../../infobox/Infobox');
 const infobox_reg = new RegExp('{{(' + i18n.infoboxes.join('|') + ')[: \n]', 'ig');
 
 //create a list of templates we can understand, and will parse later
-let keep = {};
-const parsers = [
-  require('../../sentence/templates/templates'),
-  require('./parsers'),
-];
-parsers.forEach((o) => {
-  Object.keys(o).forEach((k) => keep[k] = true);
-});
+let inlineTemplates = require('../../sentence/templates/templates');
+const parsers = require('./parsers');
+
+//templates are usually '{{name|stuff}}'
+const templateName = function(tmpl) {
+  let name = tmpl.match(/^\{\{([^:|\n ]+)/);
+  if (name === null || !name[1]) {
+    return null;
+  }
+  name = name[1].trim().toLowerCase();
+  return name;
+};
 
 //reduce the scary recursive situations
 const findTemplates = function(r, wiki, options) {
-  //remove {{template {{}} }} recursions
-  let matches = findRecursive('{', '}', wiki).filter(s => s[0] && s[1] && s[0] === '{' && s[1] === '{');
+  //grab {{template {{}} }} recursions
+  let matches = findRecursive('{', '}', wiki);
+  matches = matches.filter(s => s[0] && s[1] && s[0] === '{' && s[1] === '{');
+  //ok, go through each one...
   matches.forEach(function(tmpl) {
 
     if (tmpl.match(infobox_reg, 'ig')) {
       if (options.infoboxes !== false) {
         let infobox = parseInfobox(tmpl);
-        infobox = new Infobox(infobox, tmpl);
-        r.infoboxes = r.infoboxes || [];
-        r.infoboxes.push(infobox);
+        // infobox = new Infobox(infobox, tmpl);
+        r.templates.push({
+          template: 'infobox',
+          data: infobox
+        });
       }
       wiki = wiki.replace(tmpl, '');
       return;
     }
 
-    //keep these ones, we'll parse them later
-    let name = tmpl.match(/^\{\{([^:|\n ]+)/);
-    if (name !== null) {
-      name = name[1].trim().toLowerCase();
+    //keep these ones, we'll parse them
+    let name = templateName(tmpl);
+    if (name === null) {
+      wiki = wiki.replace(tmpl, '');
+      return;
+    }
 
-      if (/^\{\{ ?citation needed/i.test(tmpl) === true) {
-        name = 'citation needed';
+    //sorta-keep nowrap template
+    if (name === 'nowrap') {
+      let inside = tmpl.match(/^\{\{nowrap *?\|(.*?)\}\}$/);
+      if (inside) {
+        wiki = wiki.replace(tmpl, inside[1]);
       }
-      //parse {{cite web ...}} (it appears every language)
-      if (name === 'cite' || name === 'citation') {
-        wiki = parseCitation(tmpl, wiki, r, options);
-        return;
+    }
+    //parse this template
+    if (parsers.hasOwnProperty(name) === true) {
+      let template = parsers[name](tmpl, options);
+      if (template !== null) {
+        r.templates.push(template);
       }
-
-      //sorta-keep nowrap template
-      if (name === 'nowrap') {
-        let inside = tmpl.match(/^\{\{nowrap *?\|(.*?)\}\}$/);
-        if (inside) {
-          wiki = wiki.replace(tmpl, inside[1]);
-        }
-      }
-      if (keep.hasOwnProperty(name) === true) {
-        return;
-      }
+      //remove it from the wiki document
+      wiki = wiki.replace(tmpl, '');
+    }
+    if (inlineTemplates.hasOwnProperty(name) === true) {
+      return;
     }
     //here: add custom parser support?