+++ tests/citation.test.js
@@ -6,7 +6,7 @@ test('basic-citation', t => {
   var str = `Emery is a vegetarian,<ref>{{cite web|title=The princess of pot|url=http://thewalrus.ca/the-princess-of-pot/}}</ref>`;
   var arr = wtf(str).citations();
   t.equal(1, arr.length, 'found-one-citation');
-  t.equal('web', arr[0].cite, 'cite web');
+  t.equal('web', arr[0].type, 'cite web');
   t.equal('The princess of pot', arr[0].title, 'title');
   t.equal('http://thewalrus.ca/the-princess-of-pot/', arr[0].url, 'url');
   t.end();
@@ -17,7 +17,7 @@ test('complex-citation', t => {
 | url=http://cool.com/?fun=cool/}}</ref>`;
   var arr = wtf(str).citations();
   t.equal(1, arr.length, 'found-one-citation');
-  t.equal('web', arr[0].cite, 'cite web');
+  t.equal('web', arr[0].type, 'cite web');
   t.equal('bar', arr[0].foo, 'foo');
   t.equal('http://cool.com/?fun=cool/', arr[0].url, 'url');
   t.end();